requirements.txt
django==4.2.1
pillow==9.5.0
django-debug-toolbar==4.0.0
django-mptt==0.14.0
psycopg2-binary==2.9.6
requests
beautifulsoup4==4.12.2
python-decouple==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf_spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
